+++ api/v1alpha1/kafkatopic_types.go
@@ -3,6 +3,7 @@
 package v1alpha1
 
 import (
+	"github.com/aiven/go-client-codegen/handler/kafkatopic"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -31,7 +32,7 @@ type KafkaTopicSpec struct {
 	Tags []KafkaTopicTag `json:"tags,omitempty"`
 
 	// Kafka topic configuration
-	Config KafkaTopicConfig `json:"config,omitempty"`
+	Config *KafkaTopicConfig `json:"config,omitempty"`
 
 	// It is a Kubernetes side deletion protections, which prevents the kafka topic
 	// from being deleted by Kubernetes. It is recommended to enable this for any production
@@ -60,83 +61,89 @@ type KafkaTopicTag struct {
 }
 
 type KafkaTopicConfig struct {
-	// cleanup.policy value
-	CleanupPolicy string `json:"cleanup_policy,omitempty"`
+	// The retention policy to use on old segments. Possible values include 'delete', 'compact', or a comma-separated list of them. The default policy ('delete') will discard old segments when their retention time or size limit has been reached. The 'compact' setting will enable log compaction on the topic.
+	CleanupPolicy kafkatopic.CleanupPolicyType `json:"cleanup_policy,omitempty"`
 
-	// compression.type value
-	CompressionType string `json:"compression_type,omitempty"`
+	// Specify the final compression type for a given topic. This configuration accepts the standard compression codecs ('gzip', 'snappy', 'lz4', 'zstd'). It additionally accepts 'uncompressed' which is equivalent to no compression; and 'producer' which means retain the original compression codec set by the producer.
+	CompressionType kafkatopic.CompressionType `json:"compression_type,omitempty"`
 
-	// delete.retention.ms value
-	DeleteRetentionMs *int64 `json:"delete_retention_ms,omitempty"`
+	// The amount of time to retain delete tombstone markers for log compacted topics. This setting also gives a bound on the time in which a consumer must complete a read if they begin from offset 0 to ensure that they get a valid snapshot of the final stage (otherwise delete tombstones may be collected before they complete their scan).
+	DeleteRetentionMs *int `json:"delete_retention_ms,omitempty"`
 
-	// file.delete.delay.ms value
-	FileDeleteDelayMs *int64 `json:"file_delete_delay_ms,omitempty"`
+	// The time to wait before deleting a file from the filesystem.
+	FileDeleteDelayMs *int `json:"file_delete_delay_ms,omitempty"`
 
-	// flush.messages value
-	FlushMessages *int64 `json:"flush_messages,omitempty"`
+	// This setting allows specifying an interval at which we will force an fsync of data written to the log. For example if this was set to 1 we would fsync after every message; if it were 5 we would fsync after every five messages. In general we recommend you not set this and use replication for durability and allow the operating system's background flush capabilities as it is more efficient.
+	FlushMessages *int `json:"flush_messages,omitempty"`
 
-	// flush.ms value
-	FlushMs *int64 `json:"flush_ms,omitempty"`
+	// This setting allows specifying a time interval at which we will force an fsync of data written to the log. For example if this was set to 1000 we would fsync after 1000 ms had passed. In general we recommend you not set this and use replication for durability and allow the operating system's background flush capabilities as it is more efficient.
+	FlushMs *int `json:"flush_ms,omitempty"`
 
-	// index.interval.bytes value
-	IndexIntervalBytes *int64 `json:"index_interval_bytes,omitempty"`
+	// This setting controls how frequently Kafka adds an index entry to its offset index. The default setting ensures that we index a message roughly every 4096 bytes. More indexing allows reads to jump closer to the exact position in the log but makes the index larger. You probably don't need to change this.
+	IndexIntervalBytes *int `json:"index_interval_bytes,omitempty"`
 
-	// local.retention.bytes value
-	LocalRetentionBytes *int64 `json:"local_retention_bytes,omitempty"`
+	// Indicates whether inkless should be enabled.
+	InklessEnable *bool `json:"inkless_enable,omitempty"`
 
-	// local.retention.ms value
-	LocalRetentionMs *int64 `json:"local_retention_ms,omitempty"`
+	// This configuration controls the maximum bytes tiered storage will retain segment files locally before it will discard old log segments to free up space. If set to -2, the limit is equal to overall retention time. If set to -1, no limit is applied but it's possible only if overall retention is also -1.
+	LocalRetentionBytes *int `json:"local_retention_bytes,omitempty"`
 
-	// max.compaction.lag.ms value
-	MaxCompactionLagMs *int64 `json:"max_compaction_lag_ms,omitempty"`
+	// This configuration controls the maximum time tiered storage will retain segment files locally before it will discard old log segments to free up space. If set to -2, the time limit is equal to overall retention time. If set to -1, no time limit is applied but it's possible only if overall retention is also -1.
+	LocalRetentionMs *int `json:"local_retention_ms,omitempty"`
 
-	// max.message.bytes value
-	MaxMessageBytes *int64 `json:"max_message_bytes,omitempty"`
+	// The maximum time a message will remain ineligible for compaction in the log. Only applicable for logs that are being compacted.
+	MaxCompactionLagMs *int `json:"max_compaction_lag_ms,omitempty"`
 
-	// message.downconversion.enable value
+	// The largest record batch size allowed by Kafka (after compression if compression is enabled). If this is increased and there are consumers older than 0.10.2, the consumers' fetch size must also be increased so that the they can fetch record batches this large. In the latest message format version, records are always grouped into batches for efficiency. In previous message format versions, uncompressed records are not grouped into batches and this limit only applies to a single record in that case.
+	MaxMessageBytes *int `json:"max_message_bytes,omitempty"`
+
+	// This configuration controls whether down-conversion of message formats is enabled to satisfy consume requests. When set to false, broker will not perform down-conversion for consumers expecting an older message format. The broker responds with UNSUPPORTED_VERSION error for consume requests from such older clients. This configuration does not apply to any message format conversion that might be required for replication to followers.
 	MessageDownconversionEnable *bool `json:"message_downconversion_enable,omitempty"`
 
-	// message.format.version value
-	MessageFormatVersion string `json:"message_format_version,omitempty"`
+	// Specify the message format version the broker will use to append messages to the logs. The value should be a valid ApiVersion. Some examples are: 0.8.2, 0.9.0.0, 0.10.0, check ApiVersion for more details. By setting a particular message format version, the user is certifying that all the existing messages on disk are smaller or equal than the specified version. Setting this value incorrectly will cause consumers with older versions to break as they will receive messages with a format that they don't understand.
+	MessageFormatVersion kafkatopic.MessageFormatVersionType `json:"message_format_version,omitempty"`
 
-	// message.timestamp.difference.max.ms value
-	MessageTimestampDifferenceMaxMs *int64 `json:"message_timestamp_difference_max_ms,omitempty"`
+	// The maximum difference allowed between the timestamp when a broker receives a message and the timestamp specified in the message. If message.timestamp.type=CreateTime, a message will be rejected if the difference in timestamp exceeds this threshold. This configuration is ignored if message.timestamp.type=LogAppendTime.
+	MessageTimestampDifferenceMaxMs *int `json:"message_timestamp_difference_max_ms,omitempty"`
 
-	// message.timestamp.type value
-	MessageTimestampType string `json:"message_timestamp_type,omitempty"`
+	// Define whether the timestamp in the message is message create time or log append time.
+	MessageTimestampType kafkatopic.MessageTimestampType `json:"message_timestamp_type,omitempty"`
 
-	// min.cleanable.dirty.ratio value
+	// This configuration controls how frequently the log compactor will attempt to clean the log (assuming log compaction is enabled). By default we will avoid cleaning a log where more than 50% of the log has been compacted. This ratio bounds the maximum space wasted in the log by duplicates (at 50% at most 50% of the log could be duplicates). A higher ratio will mean fewer, more efficient cleanings but will mean more wasted space in the log. If the max.compaction.lag.ms or the min.compaction.lag.ms configurations are also specified, then the log compactor considers the log to be eligible for compaction as soon as either: (i) the dirty ratio threshold has been met and the log has had dirty (uncompacted) records for at least the min.compaction.lag.ms duration, or (ii) if the log has had dirty (uncompacted) records for at most the max.compaction.lag.ms period.
 	MinCleanableDirtyRatio *float64 `json:"min_cleanable_dirty_ratio,omitempty"`
 
-	// min.compaction.lag.ms value
-	MinCompactionLagMs *int64 `json:"min_compaction_lag_ms,omitempty"`
+	// The minimum time a message will remain uncompacted in the log. Only applicable for logs that are being compacted.
+	MinCompactionLagMs *int `json:"min_compaction_lag_ms,omitempty"`
 
-	// min.insync.replicas value
-	MinInsyncReplicas *int64 `json:"min_insync_replicas,omitempty"`
+	// When a producer sets acks to 'all' (or '-1'), this configuration specifies the minimum number of replicas that must acknowledge a write for the write to be considered successful. If this minimum cannot be met, then the producer will raise an exception (either NotEnoughReplicas or NotEnoughReplicasAfterAppend). When used together, min.insync.replicas and acks allow you to enforce greater durability guarantees. A typical scenario would be to create a topic with a replication factor of 3, set min.insync.replicas to 2, and produce with acks of 'all'. This will ensure that the producer raises an exception if a majority of replicas do not receive a write.
+	MinInsyncReplicas *int `json:"min_insync_replicas,omitempty"`
 
-	// preallocate value
+	// True if we should preallocate the file on disk when creating a new log segment.
 	Preallocate *bool `json:"preallocate,omitempty"`
 
-	// remote_storage_enable
+	// Indicates whether tiered storage should be enabled.
 	RemoteStorageEnable *bool `json:"remote_storage_enable,omitempty"`
 
-	// retention.bytes value
-	RetentionBytes *int64 `json:"retention_bytes,omitempty"`
+	// This configuration controls the maximum size a partition (which consists of log segments) can grow to before we will discard old log segments to free up space if we are using the 'delete' retention policy. By default there is no size limit only a time limit. Since this limit is enforced at the partition level, multiply it by the number of partitions to compute the topic retention in bytes.
+	RetentionBytes *int `json:"retention_bytes,omitempty"`
 
-	// retention.ms value
-	RetentionMs *int64 `json:"retention_ms,omitempty"`
+	// This configuration controls the maximum time we will retain a log before we will discard old log segments to free up space if we are using the 'delete' retention policy. This represents an SLA on how soon consumers must read their data. If set to -1, no time limit is applied.
+	RetentionMs *int `json:"retention_ms,omitempty"`
 
-	// segment.bytes value
-	SegmentBytes *int64 `json:"segment_bytes,omitempty"`
+	// This configuration controls the segment file size for the log. Retention and cleaning is always done a file at a time so a larger segment size means fewer files but less granular control over retention. Setting this to a very low value has consequences, and the Aiven management plane ignores values less than 10 megabytes.
+	SegmentBytes *int `json:"segment_bytes,omitempty"`
 
-	// segment.index.bytes value
-	SegmentIndexBytes *int64 `json:"segment_index_bytes,omitempty"`
+	// This configuration controls the size of the index that maps offsets to file positions. We preallocate this index file and shrink it only after log rolls. You generally should not need to change this setting.
+	SegmentIndexBytes *int `json:"segment_index_bytes,omitempty"`
 
-	// segment.jitter.ms value
-	SegmentJitterMs *int64 `json:"segment_jitter_ms,omitempty"`
+	// The maximum random jitter subtracted from the scheduled segment roll time to avoid thundering herds of segment rolling
+	SegmentJitterMs *int `json:"segment_jitter_ms,omitempty"`
 
-	// segment.ms value
-	SegmentMs *int64 `json:"segment_ms,omitempty"`
+	// This configuration controls the period of time after which Kafka will force the log to roll even if the segment file isn't full to ensure that retention can delete or compact old data. Setting this to a very low value has consequences, and the Aiven management plane ignores values less than 10 seconds.
+	SegmentMs *int `json:"segment_ms,omitempty"`
+
+	// Indicates whether to enable replicas not in the ISR set to be elected as leader as a last resort, even though doing so may result in data loss.
+	UncleanLeaderElectionEnable *bool `json:"unclean_leader_election_enable,omitempty"`
 }
 
 // KafkaTopicStatus defines the observed state of KafkaTopic
@@ -145,7 +152,7 @@ type KafkaTopicStatus struct {
 	Conditions []metav1.Condition `json:"conditions"`
 
 	// State represents the state of the kafka topic
-	State string `json:"state"`
+	State kafkatopic.TopicStateType `json:"state"`
 }
 
 // +kubebuilder:object:root=true