+++ tests/kafkatopic_test.go
@@ -116,50 +116,48 @@ func TestKafkaTopic(t *testing.T) {
 	assert.True(t, meta.IsStatusConditionTrue(barTopic.Status.Conditions, "Running"))
 
 	// KafkaTopic with name `foo-topic`
-	fooAvn, err := avnClient.KafkaTopics.Get(ctx, cfg.Project, ksName, fooTopic.GetTopicName())
+	fooAvn, err := avnGen.ServiceKafkaTopicGet(ctx, cfg.Project, ksName, fooTopic.GetTopicName())
 	require.NoError(t, err)
 	assert.Equal(t, "foo-topic", fooTopic.GetName())
 	assert.Equal(t, "foo-topic", fooTopic.GetTopicName())
 	assert.Equal(t, fooAvn.TopicName, fooTopic.GetTopicName())
-	assert.Equal(t, "ACTIVE", fooTopic.Status.State)
+	assert.Equal(t, fooAvn.State, fooTopic.Status.State)
 	assert.Equal(t, fooAvn.State, fooTopic.Status.State)
 	assert.Equal(t, fooAvn.Replication, fooTopic.Spec.Replication)
 	assert.Len(t, fooAvn.Partitions, fooTopic.Spec.Partitions)
 
 	// Validates MinCleanableDirtyRatio
 	require.Equal(t, anyPointer(0.2), fooTopic.Spec.Config.MinCleanableDirtyRatio)
-	require.Equal(t, anyPointer(int64(2048)), fooTopic.Spec.Config.RetentionBytes)
-	require.Equal(t, anyPointer(int64(1024)), fooTopic.Spec.Config.LocalRetentionBytes)
-	require.Equal(t, anyPointer(int64(1000000)), fooTopic.Spec.Config.LocalRetentionMs)
+	require.Equal(t, anyPointer(2048), fooTopic.Spec.Config.RetentionBytes)
+	require.Equal(t, anyPointer(1024), fooTopic.Spec.Config.LocalRetentionBytes)
+	require.Equal(t, anyPointer(1000000), fooTopic.Spec.Config.LocalRetentionMs)
+
+	// Validates MaxMessageBytes (not set)
+	assert.Nil(t, fooTopic.Spec.Config.MaxMessageBytes)
 
 	// KafkaTopic with name `bar_topic_name_with_underscores`
-	barAvn, err := avnClient.KafkaTopics.Get(ctx, cfg.Project, ksName, barTopic.GetTopicName())
+	barAvn, err := avnGen.ServiceKafkaTopicGet(ctx, cfg.Project, ksName, barTopic.GetTopicName())
 	require.NoError(t, err)
 	assert.Equal(t, "bar-topic", barTopic.GetName())
 	assert.Equal(t, "bar_topic_name_with_underscores", barTopic.GetTopicName())
 	assert.Equal(t, barAvn.TopicName, barTopic.GetTopicName())
-	assert.Equal(t, "ACTIVE", barTopic.Status.State)
+	assert.Equal(t, barAvn.State, barTopic.Status.State)
 	assert.Equal(t, barAvn.State, barTopic.Status.State)
 	assert.Equal(t, barAvn.Replication, barTopic.Spec.Replication)
 	assert.Len(t, barAvn.Partitions, barTopic.Spec.Partitions)
-
-	// Validates MinCleanableDirtyRatio (not set)
-	assert.Nil(t, barTopic.Spec.Config.MinCleanableDirtyRatio)
-	assert.Nil(t, barTopic.Spec.Config.LocalRetentionBytes)
-	assert.Nil(t, barTopic.Spec.Config.LocalRetentionMs)
-	assert.Nil(t, barTopic.Spec.Config.RemoteStorageEnable)
+	assert.Nil(t, barTopic.Spec.Config)
 
 	// We need to validate deletion,
 	// because we can get false positive here:
 	// if service is deleted, topic is destroyed in Aiven. No service — no topic. No topic — no topic.
 	// And we make sure that controller can delete topic itself
 	assert.NoError(t, s.Delete(fooTopic, func() error {
-		_, err = avnClient.KafkaTopics.Get(ctx, cfg.Project, ksName, fooTopic.Name)
+		_, err = avnGen.ServiceKafkaTopicGet(ctx, cfg.Project, ksName, fooTopic.Name)
 		return err
 	}))
 
 	assert.NoError(t, s.Delete(barTopic, func() error {
-		_, err = avnClient.KafkaTopics.Get(ctx, cfg.Project, ksName, "bar_topic_name_with_underscores")
+		_, err = avnGen.ServiceKafkaTopicGet(ctx, cfg.Project, ksName, "bar_topic_name_with_underscores")
 		return err
 	}))
 }